+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.pddstudio.highlightjs.demo.MainActivity">
+    tools:context="com.medium.highlightjs.demo.MainActivity">
 
     <RelativeLayout
         android:id="@+id/activity_main"
