+++ app/src/main/res/layout/activity_syntax.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_syntax"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.pddstudio.highlightjs.demo.SyntaxActivity">
+    tools:context="com.medium.highlightjs.demo.SyntaxActivity">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swipe_refresh_layout"
@@ -16,7 +16,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent">
 
-            <com.pddstudio.highlightjs.HighlightJsView
+            <com.medium.highlightjs.HighlightJsView
                 android:id="@+id/highlight_view"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
