--- app/src/test/java/com/pddstudio/highlightjs/demo/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package com.pddstudio.highlightjs.demo;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
