--- library/src/androidTest/java/com/pddstudio/highlightjs/ExampleInstrumentationTest.java
@@ -1,29 +1,0 @@
-package com.pddstudio.highlightjs;
-
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.filters.MediumTest;
-import android.support.test.runner.AndroidJUnit4;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-
-import static org.junit.Assert.*;
-
-/**
- * Instrumentation test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@MediumTest
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentationTest {
-    @Test
-    public void useAppContext() throws Exception {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
-
-        assertEquals("com.pddstudio.highlightjs.test", appContext.getPackageName());
-    }
-}
