+++ library/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.pddstudio.highlightjs" />
+    package="com.medium.highlightjs" />
