+++ library/src/main/java/com/medium/highlightjs/models/SelectionCallback.kt
@@ -1,0 +1,5 @@
+package com.medium.highlightjs.models
+
+interface SelectionCallback {
+    fun onSelectionChange(selectedText: String?)
+}
