--- library/src/main/java/com/pddstudio/highlightjs/models/SelectionCallback.kt
@@ -1,5 +1,0 @@
-package com.pddstudio.highlightjs.models
-
-interface SelectionCallback {
-    fun onSelectionChange(selectedText: String?)
-}
